Move `src` -> `lib`.
Using `src` in `dist/` was somewhat annoying, and it also made things
ever so slightly confusing (is this an application/addon in the MU
world?).

This is basically a straight rename via `git mv src lib`.
