+++ lib/broccoli/default-module-configuration.ts
