+++ lib/broccoli/glimmer-template-precompiler.ts
