+++ lib/broccoli/rollup-with-dependencies.ts
