+++ testem.json
@@ -1,7 +1,7 @@
 {
   "framework": "qunit",
   "test_page": "index.html",
-  "src_files": ["src/**/*", "test/**/*"],
+  "src_files": ["lib/**/*", "test/**/*"],
   "disable_watching": true,
   "launch_in_ci": ["phantomjs"]
 }
